requirements.txt
flask==3.0.0
geoip2==4.7.0
pyyaml==6.0.1
gunicorn==21.2.0
requests==2.31.0
